Move block production specific stuff to block_production module
